+++ src/services/no-op.ts
@@ -1,0 +1,22 @@
+import { Composable } from '../models/composable.js';
+import { EventEnvelope } from '../models/event-envelope.js';
+import { preload } from '../models/composable.js'
+
+export class NoOp implements Composable {
+    
+    name = "no.op";
+
+    @preload()
+    initialize(): void {
+        // no-op
+    }
+
+    getName(): string {
+        return this.name;
+    }
+
+    async handleEvent(evt: EventEnvelope) {
+       return evt;
+    }
+
+}