+++ CHANGELOG.md
@@ -6,6 +6,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 4.1.1, 12/22/2024
+
+### Added
+
+1. Composable class scanner for the source folder
+2. Added "web.component.scan" parameter to support scanning of dependency libaries
+
+### Removed
+
+N/A
+
+### Changed
+
+N/A
+
+---
 ## Version 4.1.0, 12/20/2024
 
 ### Added